fix : code program use sigle thread instead multi thread for time reduncing

+++ include/IO.hpp
@@ -16,7 +16,7 @@ std::string directionsToString(const std::vector<char>& path);
 void writeBoardWithActor(std::ostream& output, const Board& board, Position actor);
 void writeSolutionSteps(std::ostream& output, const Board& board, const std::vector<char>& path, const std::vector<Position>& positions);
 void saveSolution(const Board& board, const std::vector<char>& path, const std::vector<Position>& positions, int totalCost,
-                int totalIterations, long long elapsedMs, const std::string& outputPath);
+                int totalIterations, double elapsedMs, const std::string& outputPath);
 void saveSolution(const Board& board, const std::vector<char>& path, const std::vector<Position>& positions, int totalCost,
-                int totalIterations, long long elapsedMs, const std::string& outputPath, const std::string& algorithm,
+                int totalIterations, double elapsedMs, const std::string& outputPath, const std::string& algorithm,
                 const std::string& heuristic);
